+++ docs/release-notes.rst
@@ -38,6 +38,7 @@ Version 1.2.0, 15 Aug 2019
 * Add controller for page refresh of Angular routes
 * Extend E2 mock configuration for demo purposes
 * Add pattern for matching AC/admin application name
+* Add custom (plain but not white-label) error page
 
 Version 1.0.5, 5 July 2019
 --------------------------